Move town hall functions to separate file
