+++ src/GUI/UI/Houses/TownHall.cpp
@@ -1,6 +1,6 @@
-#include "GUIBountyHunting.h"
-
 #include <cassert>
+
+#include "GUI/UI/Houses/TownHall.h"
 
 #include "Engine/Graphics/IRender.h"
 #include "Engine/Localization.h"
@@ -8,11 +8,131 @@
 
 #include "GUI/UI/UIHouses.h"
 
-#include "GUIWindow.h"
-#include "GUIButton.h"
+#include "GUI/GUIWindow.h"
+#include "GUI/GUIButton.h"
+#include "GUI/GUIFont.h"
 
 static const char *bountyHunting_text = nullptr;                // word_F8B1A4
 static int16_t bountyHunting_monsterId = 0;
+
+//----- (004B7911) --------------------------------------------------------
+void TownHallDialog() {
+    int v1;                       // eax@10
+    int v2;                       // esi@10
+    signed int pStringSum;        // ebx@24
+    signed int v16;               // ebx@28
+    int v17;                      // ebx@28
+    GUIButton *pButton;           // eax@30
+    int pTextHeight;              // eax@30
+    Color pTextColor;  // ax@30
+    int v29;                      // [sp+10Ch] [bp-10h]@28
+    int v31;                      // [sp+114h] [bp-8h]@29
+    GUIFont *pOutString;          // [sp+118h] [bp-4h]@21
+
+    GUIWindow townHall_window = *window_SpeakInHouse;
+    townHall_window.uFrameX = SIDE_TEXT_BOX_POS_X;
+    townHall_window.uFrameWidth = SIDE_TEXT_BOX_WIDTH;
+    townHall_window.uFrameZ = SIDE_TEXT_BOX_POS_Z;
+
+    std::string fine_str = fmt::format("{}: {}", localization->GetString(LSTR_CURRENT_FINE), pParty->uFine);
+    townHall_window.DrawTitleText(pFontArrus, 0, 260, colorTable.PaleCanary, fine_str, 3);
+
+    switch (dialog_menu_id) {
+    case DIALOGUE_MAIN:
+    {
+        pStringSum = 1;
+        pTextHeight = 0;
+        pShopOptions[0] = localization->GetString(LSTR_BOUNTY_HUNT);
+        if (pParty->uFine > 0) {
+            pShopOptions[1] = localization->GetString(LSTR_PAY_FINE);
+            pStringSum = 2;
+        }
+        for (uint i = 0; i < pStringSum; ++i)
+            pTextHeight += pFontArrus->CalcTextHeight(
+                pShopOptions[i], townHall_window.uFrameWidth, 0);
+        v29 = (100 - pTextHeight) / pStringSum;
+        v16 = 80 - pStringSum * ((100 - pTextHeight) / pStringSum);
+        v17 = (v16 / 2) - v29 / 2 + 158;
+        if (pDialogueWindow->pNumPresenceButton > 0) {
+            v31 = 2;
+            uint j = 0;
+            for (uint i = pDialogueWindow->pStartingPosActiveItem;
+                i < pDialogueWindow->pNumPresenceButton +
+                pDialogueWindow->pStartingPosActiveItem;
+                ++i) {
+                pButton = pDialogueWindow->GetControl(i);
+                pButton->uY = v29 + v17;
+                pTextHeight = pFontArrus->CalcTextHeight(
+                    pShopOptions[j], townHall_window.uFrameWidth, 0);
+                pButton->uHeight = pTextHeight;
+                v17 = pButton->uY + pTextHeight - 1;
+                pButton->uW = v17 + 6;
+                pTextColor = colorTable.PaleCanary;
+                if (pDialogueWindow->pCurrentPosActiveItem != v31)
+                    pTextColor = colorTable.White;
+                townHall_window.DrawTitleText(pFontArrus, 0, pButton->uY, pTextColor, pShopOptions[j], 3);
+                ++v31;
+                ++j;
+            }
+        }
+        break;
+    }
+    case DIALOGUE_TOWNHALL_MESSAGE:
+    {
+        current_npc_text = bountyHuntingText();
+        GUIWindow window = *pDialogueWindow;
+        window.uFrameWidth = 458;
+        window.uFrameZ = 457;
+        pOutString = pFontArrus;
+        pTextHeight = pFontArrus->CalcTextHeight(current_npc_text, window.uFrameWidth, 13) + 7;
+        if (352 - pTextHeight < 8) {
+            pOutString = pFontCreate;
+            pTextHeight = pFontCreate->CalcTextHeight(current_npc_text, window.uFrameWidth, 13) + 7;
+        }
+        render->DrawTextureCustomHeight(8 / 640.0f, (352 - pTextHeight) / 480.0f, ui_leather_mm7, pTextHeight);
+        render->DrawTextureNew(8 / 640.0f, (347 - pTextHeight) / 480.0f, _591428_endcap);
+        // window.DrawText(pOutString, 13, 354 - pTextHeight, 0, pOutString->FitTextInAWindow(current_npc_text, window.uFrameWidth, 13), 0, 0, 0);
+        window.DrawText(pOutString, {13, 354 - pTextHeight}, Color(), current_npc_text, 0, 0, Color());
+        break;
+    }
+    case DIALOGUE_TOWNHALL_PAY_FINE:
+    {
+        if (window_SpeakInHouse->keyboard_input_status == WINDOW_INPUT_IN_PROGRESS) {
+            townHall_window.DrawTitleText(pFontArrus, 0, 146, colorTable.PaleCanary,
+                                          fmt::format("{}\n{}", localization->GetString(LSTR_PAY), localization->GetString(LSTR_HOW_MUCH)), 3);
+            townHall_window.DrawTitleText(pFontArrus, 0, 186, colorTable.White, keyboardInputHandler->GetTextInput(), 3);
+            townHall_window.DrawFlashingInputCursor(pFontArrus->GetLineWidth(keyboardInputHandler->GetTextInput()) / 2 + 80, 185, pFontArrus);
+            return;
+        } else if (window_SpeakInHouse->keyboard_input_status == WINDOW_INPUT_CONFIRMED) {
+            int sum = atoi(keyboardInputHandler->GetTextInput().c_str());
+            if (sum > 0) {
+                int party_gold = pParty->GetGold();
+                if (sum > party_gold) {
+                    PlayHouseSound(window_SpeakInHouse->wData.val, HouseSound_NotEnoughMoney);
+                    sum = party_gold;
+                }
+
+                if (sum > 0) {
+                    int required_sum = pParty->GetFine();
+                    if (sum > required_sum)
+                        sum = required_sum;
+
+                    pParty->TakeGold(sum);
+                    pParty->TakeFine(sum);
+                    if (pParty->hasActiveCharacter())
+                        pParty->activeCharacter().playReaction(SPEECH_BankDeposit);
+                }
+            }
+        }
+        window_SpeakInHouse->keyboard_input_status = WINDOW_INPUT_NONE;
+        pCurrentFrameMessageQueue->AddGUIMessage(UIMSG_Escape, 1, 0);
+        break;
+    }
+    default:
+        break;
+    }
+    return;
+}
 
 static int RandomMonsterForHunting(HOUSE_ID townhall) {
     while (true) {
@@ -116,20 +236,6 @@ static int RandomMonsterForHunting(HOUSE_ID townhall) {
             return -1;
         }
     }
-}
-
-void openBountyHuntingDialogue() {
-    uDialogueType = DIALOGUE_83_bounty_hunting;
-    pDialogueWindow->Release();
-    pDialogueWindow = new GUIWindow(WINDOW_Dialogue, {0, 0}, {render->GetRenderDimensions().w, 350}, 0);
-    pBtn_ExitCancel = pDialogueWindow->CreateButton({471, 445}, {169, 35}, 1, 0, UIMSG_Escape, 0, InputAction::Invalid,
-                                                    localization->GetString(LSTR_CANCEL), { ui_exit_cancel_button_background });
-    pDialogueWindow->CreateButton({0, 0}, {0, 0}, 1, 0, UIMSG_BuyInShop_Identify_Repair, 0, InputAction::Invalid, "");
-    pDialogueWindow->CreateButton({480, 160}, {140, 30}, 1, 0, UIMSG_0, DIALOGUE_83_bounty_hunting, InputAction::Invalid, "");
-    pDialogueWindow->_41D08F_set_keyboard_control_group(1, 1, 0, 2);
-    dialog_menu_id = DIALOGUE_OTHER;
-
-    bountyHuntingDialogueOptionClicked();
 }
 
 void bountyHuntingDialogueOptionClicked() {
